Remove Babel SVG plugin (#128)
This PR removes the need for the `babel-plugin-inline-react-svg` plugin. The plugin was previously used to transform SVG imports into React components that could be used on a page.

There are a few benefits to removing this:

1. The babel plugin (or Gatsby's babel setup?) is caching SVGs too aggressively. Changes to SVGs are not appearing until we restarted Gatsby. This caused a bug two weeks ago when I minified the star in the `StarRating` component and didn't realize that I had broken it.
2. This means we now have no custom babel plugins. Simplifies config and makes it easier for developers to work in the Thumbprint codebase.
3. Importing images in Gatsby will now be consistent across SVGs and non-SVGs.

There are two types of JSX files that were changed:

* `packages/**/*.jsx`: I've converted the SVGs into tiny React components and minimized the SVG.
* `www/**/*.jsx`: Gatsby supports importing SVGs out-of-the-box. The imports now return a URL instead of a React component.

This modifies three React components: `StarRating`, `ModalStandard`, and `Avatar`. I've updated snapshots and manually verified that they are rendering correctly.