Create codemods package. (#183)
This is a port from code in a separate repo. By putting it here and publishing to NPM, we can make it easier to use because consumers will be able to do `npx @thumbtack/thumbprint-codemods` and use the CLI without needing to clone a repo.

It also allows us to benefit from the lint, testing, and pretter infra that this codebase already has.