Improve cache that determines when to build packages (#225)
We have a basic cache that determines when to build packages. The cache previously worked by creating a hash of the package folder and only rebuilding when the package folder changed. This means that changes to global (all of Thumbprint) dependencies would not force packages to re-build. This led to mismatches between CI and local because since Jenkins always starts from scratch and never uses the cache.

This fixes #212, an issue where a version bump of browserslist was failing locally but passing in CI. This commit fixes it because changes to `yarn.lock` will now invalidate the cache.